requirements.txt
Django>=5.1,<5.2
python-decouple>=3.8,<3.9
psycopg2>=2.9.10,<2.10
djangorestframework>=3.15.1,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4